Call success callback after checking connection
Prior to this change I've noticed that connect will call the success
callback before the connection is actually established. This will
ensure that the device is connected to the correct network and that the
state of the connection is one where network calls can be made right
after the success callback returns.